Align with EventStore-5.0.8

+++ src/EventStore.Projections.Core/Services/Processing/CheckpointTag.cs
@@ -702,8 +702,8 @@ namespace EventStore.Projections.Core.Services.Processing
             long? preparePosition = null;
             string catalogStream = null;
             string dataStream = null;
-            int? catalogPosition = null;
-            int? dataPosition = null;
+            long? catalogPosition = null;
+            long? dataPosition = null;
             bool byStreamMode = false;
             Dictionary<string, long> streams = null;
             Dictionary<string, JToken> extra = null;
@@ -779,7 +779,7 @@ namespace EventStore.Projections.Core.Services.Processing
                             Check(JsonToken.PropertyName, reader);
                             catalogStream = (string)reader.Value;
                             Check(reader.Read(), reader);
-                            catalogPosition = (int)(long)reader.Value;
+                            catalogPosition = (long)reader.Value;
                             Check(reader.Read(), reader);
                             Check(JsonToken.EndObject, reader);
 
@@ -790,7 +790,7 @@ namespace EventStore.Projections.Core.Services.Processing
                                 Check(JsonToken.PropertyName, reader);
                                 dataStream = (string)reader.Value;
                                 Check(reader.Read(), reader);
-                                dataPosition = (int)(long)reader.Value;
+                                dataPosition = (long)reader.Value;
                                 Check(reader.Read(), reader);
                                 Check(JsonToken.EndObject, reader);
                                 Check(reader.Read(), reader);
@@ -808,7 +808,7 @@ namespace EventStore.Projections.Core.Services.Processing
                                 Check(JsonToken.PropertyName, reader);
                                 var streamName = (string)reader.Value;
                                 Check(reader.Read(), reader);
-                                var position = (int)(long)reader.Value;
+                                var position = (long)reader.Value;
                                 streams.Add(streamName, position);
                             }
                         }
